Move test_helpers.go to apptesting folder. Remove unnecessary EncodingConfig struct/method/file. Move and rename mock.go to mock_key.go and move to apptesting folder.
